test: extract MockRandomNumberGenerator class
