+++ src/test/java/racingcar/controller/RacingGameControllerTest.java
@@ -3,15 +3,14 @@ package racingcar.controller;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import java.util.List;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import racingcar.controller.exception.RacingGameIsFinishedException;
 import racingcar.domain.Car;
 import racingcar.domain.MidtermResult;
+import racingcar.domain.MockRandomNumberGenerator;
 import racingcar.repository.CarRepository;
 import racingcar.service.RacingGameService;
-import racingcar.service.RacingGameServiceTest;
 
 class RacingGameControllerTest {
 
@@ -22,14 +21,15 @@ class RacingGameControllerTest {
     @BeforeEach
     void setUp() {
         carRepository = new CarRepository();
-        racingGameService = new RacingGameService(carRepository, new RacingGameServiceTest.MockRandomNumberGenerator());
+        racingGameService = new RacingGameService(carRepository, new MockRandomNumberGenerator());
         racingGameController = new RacingGameController(carRepository, racingGameService);
     }
 
     @Test
     public void 자동차_이름_입력() {
         racingGameController.inputCarNames("pobi,crong,honux");
-        assertThat(carRepository.findAll()).contains(Car.from("pobi"), Car.from("crong"), Car.from("honux"));
+        assertThat(carRepository.findAll())
+            .contains(Car.from("pobi"), Car.from("crong"), Car.from("honux"));
     }
 
     @Test
